docs(release): defer headless core to roadmap
