+++ .changeset/calm-input-hosts.md
@@ -1,0 +1,6 @@
+---
+'@asyra/input-system': patch
+---
+
+Make InputSystem import and construction DOM-neutral, add explicit browser host
+and pointer-target lifecycle APIs, and provide symmetric listener cleanup.
