+++ docs/ai/framework/decisions/releases/unreleased.md
@@ -2346,3 +2346,39 @@ unregister -> app migration -> core.start()` as the public app route.
   - [#113](https://github.com/karote00/asyra/pull/113)
 - Accepted Source Commit:
   - `0bbb1cc35dc3cb1eebae30b7c81c494afbd063f8`
+
+## 2026-08-10 - Defer Headless Core and Core Kernel; retain Input environment neutrality
+
+- Context:
+  - Input System eagerly read `window` and registered listeners during
+    construction, preventing otherwise environment-neutral public imports.
+  - A proposed Headless Core factory exposed larger unresolved ownership issues:
+    package-default subscribers, process-scoped registries and Feature state,
+    and transaction replay owner routing could create false instance isolation.
+  - The public-release website plans were beginning to treat non-visible runtime
+    behavior as a current executable contract.
+- Decision:
+  - Keep the release-scoped Input System correction: import/construction is
+    DOM-neutral, and browser host/target listener ownership is explicit,
+    idempotent, transferable, and disposable.
+  - Preserve Core's existing typed watched-element event route and require
+    direct product-owner visual input testing before merge.
+  - Do not add `createHeadlessCore()`, `startHeadless()`, or a Core Kernel in the
+    initial release.
+  - Move Headless Core/Core Kernel to an unscheduled future plan and retain the
+    architecture research report as its indexed starting evidence.
+- Consequences:
+  - Node-safe Input/Core import is documented only as environment safety, not as
+    a non-visible runtime, no-Render/UI dependency, or isolation promise.
+  - Website, public docs, examples, and Runtime Atlas may explain the future
+    direction but cannot present Headless/Core Kernel behavior as current
+    support.
+  - Future activation requires a fresh current-code audit, explicit capability
+    decision, bounded plan, exact Inspector, test-first owner/isolation proofs,
+    and clean-consumer evidence.
+- Active Plan:
+  - `docs/ai/framework/plans/input-system-environment-neutrality-plan.md`
+- Future Plan:
+  - `docs/ai/framework/plans/headless-core-and-core-kernel-future-plan.md`
+- Research Report:
+  - `docs/ai/framework/research/headless-core-and-core-kernel-architecture-research.md`
