+++ docs/ai/framework/packages/input-system.md
@@ -9,6 +9,7 @@ Normalize raw keyboard/mouse/pointer input into framework input events.
 - key-combination mapping
 - pointer event normalization (client/workspace conversion hooks if configured)
 - event dispatch to runtime consumers
+- explicit, instance-owned browser keyboard and pointer listener lifecycle
 
 ## Must Not Own
 
@@ -30,12 +31,27 @@ Normalize raw keyboard/mouse/pointer input into framework input events.
   the same callback with `off(event, callback)`. `off` removes only the requested
   listener, returns `false` when it was not registered, and removes an empty
   event bucket.
+- Import and construction must not read `window`/`document` or attach browser
+  listeners.
+- `attachBrowserHost(host, pointerTarget?)` is the explicit browser activation
+  boundary. Keyboard listeners belong to `host`; pointer/wheel listeners belong
+  to `pointerTarget`, defaulting to `host`. Repeating the same pair is
+  idempotent.
+- `switchWatchedElement(element)` derives the element's owning `Window`. A
+  target/document switch removes exact prior listeners before attaching the new
+  host/target pair.
+- `detachBrowserHost()` removes all browser listeners owned by the instance.
+  `reset()` preserves the active attachment while clearing transient state;
+  `dispose()` detaches and clears transient state.
+- Core visual startup retains the typed watched-element event route. DOM-neutral
+  import/construction is not a public Headless Core or Core Kernel guarantee.
 
 ## Extension Points
 
 - key-combination registration
 - input event mapping per app/runtime mode
 - optional app-level adapters for host environment differences
+- explicit browser host and pointer-target selection
 
 ## Facade Note
 
@@ -47,3 +63,7 @@ Normalize raw keyboard/mouse/pointer input into framework input events.
 - Pointer down/move/up sequence remains ordered and complete.
 - Input event payloads match typed contracts.
 - Listener cleanup does not remove other consumers of the same normalized event.
+- Node import/construction succeeds with no browser globals.
+- Identical attachment is idempotent; same-document and cross-document switches
+  remove the exact previous listeners.
+- Reset preserves attachment and dispose removes it.
