+++ docs/ai/framework/plans/input-system-environment-neutrality-flow-inspector.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>Headless Core Runtime Flow Inspector</title>
+    <title>Input System Environment Neutrality Flow Inspector</title>
     <style>
       :root {
         color-scheme: dark;
@@ -98,7 +98,7 @@
       </main>
       <aside id="detail"></aside>
     </div>
-    <script src="./headless-core-runtime-flow-inspector.data.cjs"></script>
+    <script src="./input-system-environment-neutrality-flow-inspector.data.cjs"></script>
     <script>
       ;(() => {
         const data = globalThis.FLOW_INSPECTOR_DATA
